refactor: Remove unused fetchCredits function and update import paths
